+++ NEWS.md
@@ -1,7 +1,7 @@
 # crew 1.3.0.9002 (development)
 
 * Use `later::global_loop()` instead of `later::current_loop()` in `controller$autoscale()`.
-* Test `crew_worker()` in a background process (https://github.com/r-lib/mirai/issues/535).
+* Test `crew_worker()` with `autoexit = FALSE` (https://github.com/r-lib/mirai/issues/535).
 
 # crew 1.3.0